layout/svg/tests/test_disabled_chrome.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=744830
-->
<head>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=166235">Mozilla Bug 166235</a>
<div id="testnodes"><span>hi</span> there <!-- mon ami --></div>
<pre id="test">
<script type="application/javascript">
  add_task(async function() {
    const initialPrefValue = SpecialPowers.getBoolPref("svg.disabled");
    SpecialPowers.setBoolPref("svg.disabled", true);
    const Cu = SpecialPowers.Components.utils;
    const { ContentTaskUtils } = ChromeUtils.importESModule(
      "resource://testing-common/ContentTaskUtils.sys.mjs"
    );
    let t = document.getElementById('testnodes');

    let url = 'chrome://mochitests/content/chrome/layout/svg/tests/svg_example_test.html'
    const chromeIframeEl = document.createElement('iframe');
    let chromeLoadPromise = ContentTaskUtils.waitForEvent(chromeIframeEl, 'load', false);
    chromeIframeEl.src = url;
    t.appendChild(chromeIframeEl);

    await chromeLoadPromise;
    const chromeBR = chromeIframeEl.contentDocument.body.getBoundingClientRect();

    url = "http://mochi.test:8888/chrome/layout/svg/tests/svg_example_test.html";
    const iframeEl = document.createElement('iframe');
    iframeEl.src = url;
    let loadPromise = ContentTaskUtils.waitForEvent(iframeEl, 'load', false);
    t.appendChild(iframeEl);
    await loadPromise;

    const contentBR = iframeEl.contentDocument.body.getBoundingClientRect();
    ok(chromeBR.height > contentBR.height, "Chrome content height should be bigger than content due to layout");

    url = "http://mochi.test:8888/chrome/layout/svg/tests/svg_example_script.svg";
    const iframeElScript = document.createElement("iframe");
    let loadPromiseScript = ContentTaskUtils.waitForEvent(iframeElScript, "load", false);
    iframeElScript.src = url;
    t.appendChild(iframeElScript);
    await loadPromiseScript;
    ok(!iframeElScript.contentDocument.documentElement.style, "Content should not be styled");

    SpecialPowers.setBoolPref("svg.disabled", initialPrefValue);
  });
</script>
</pre>
</body>
</html>
